# docs/api/openapi-completo.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Proyecto Arquitectura - API Unificada",
    "description": "API Gateway que expone los microservicios del ecosistema de e-commerce. Incluye servicios de gestion de ordenes (Order Service), catalogo de productos (Catalog Service) con cache Redis, y busqueda full-text (Search) con Elasticsearch.\n\n## Arquitectura\n- **API Gateway**: Punto de entrada centralizado (Spring Cloud Gateway)\n- **Order Service**: Gestion de ordenes con notificaciones via SQS/Lambda\n- **Catalog Service**: CRUD de productos con cache Redis\n- **Search Service**: Busqueda full-text con Elasticsearch\n\n## Tecnologias\n- Spring Boot 3.x\n- PostgreSQL (persistencia)\n- Redis (cache)\n- Elasticsearch (busqueda)\n- AWS SQS + Lambda (mensajeria serverless)\n- Docker (contenedores)",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo de Desarrollo",
      "email": "dev@proyecto.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "API Gateway (produccion local)"
    },
    {
      "url": "http://localhost:8081",
      "description": "Order Service (desarrollo directo)"
    },
    {
      "url": "http://localhost:8082",
      "description": "Catalog Service (desarrollo directo)"
    }
  ],
...
  "paths": {
...
    "/catalog/products": {
      "get": {
        "tags": ["Catalog"],
        "summary": "Obtener todos los productos",
        "description": "Retorna la lista completa de productos. Los resultados se cachean en Redis para mejorar el rendimiento.",
        "operationId": "getAllProducts",
        "responses": {
          "200": {
            "description": "Lista de productos obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "name": "Laptop HP Pavilion",
                    "price": 999.99,
                    "stock": 50
                  },
                  {
                    "id": 2,
                    "name": "Mouse Logitech MX Master",
                    "price": 99.99,
                    "stock": 150
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Catalog"],
        "summary": "Crear nuevo producto",
        "description": "Crea un nuevo producto en el catalogo. Invalida el cache de productos en Redis.",
        "operationId": "createProduct",
        "requestBody": {
          "description": "Datos del producto a crear",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              },
              "example": {
                "name": "Teclado Mecanico RGB",
                "price": 149.99,
                "stock": 75
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Datos de producto invalidos"
          }
        }
      }
    },
    "/catalog/products/{id}": {
      "get": {
        "tags": ["Catalog"],
        "summary": "Obtener producto por ID",
        "description": "Busca y retorna un producto especifico. El resultado se cachea en Redis.",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      },
      "delete": {
        "tags": ["Catalog"],
        "summary": "Eliminar producto",
        "description": "Elimina un producto del catalogo. Invalida el cache de productos en Redis.",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto a eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Producto eliminado exitosamente"
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    },
    "/catalog/products/{id}/stock": {
      "put": {
        "tags": ["Catalog"],
        "summary": "Actualizar stock de producto",
        "description": "Actualiza la cantidad de stock de un producto. Invalida el cache de productos en Redis.",
        "operationId": "updateProductStock",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "description": "Nueva cantidad de stock",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "minimum": 0,
                "example": 100
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stock actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado"
          }
        }
      }
    },
    "/catalog/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Buscar productos",
        "description": "Busca productos por nombre, estado de stock o precio maximo usando Elasticsearch. Si no se proporcionan parametros, retorna todos los productos indexados.",
        "operationId": "searchProducts",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Texto a buscar en el nombre del producto (busqueda full-text)",
            "required": false,
            "schema": {
              "type": "string",
              "example": "laptop"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Estado de stock del producto",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["OK", "LOW", "OUT_OF_STOCK"],
              "example": "OK"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "description": "Precio maximo del producto",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "example": 500.00
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Busqueda realizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDocument"
                  }
                },
                "example": [
                  {
                    "id": "1",
                    "name": "Laptop HP Pavilion",
                    "price": 999.99,
                    "stock": 50,
                    "stockStatus": "OK"
                  }
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "description": "Producto del catalogo",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "ID unico del producto",
            "readOnly": true,
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Laptop HP Pavilion"
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Precio del producto en dolares",
            "minimum": 0,
            "example": 999.99
          },
          "stock": {
            "type": "integer",
            "description": "Cantidad disponible en stock",
            "minimum": 0,
            "example": 50
          }
        },
        "required": ["name", "price", "stock"]
      },
      "ProductCreate": {
        "type": "object",
        "description": "Datos para crear un nuevo producto",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Teclado Mecanico RGB"
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Precio del producto en dolares",
            "minimum": 0,
            "example": 149.99
          },
          "stock": {
            "type": "integer",
            "description": "Cantidad inicial en stock",
            "minimum": 0,
            "example": 75
          }
        },
        "required": ["name", "price", "stock"]
      },
      "ProductDocument": {
        "type": "object",
        "description": "Documento de producto indexado en Elasticsearch",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID del producto",
            "example": "1"
          },
          "name": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Laptop HP Pavilion"
          },
          "price": {
            "type": "number",
            "format": "double",
            "description": "Precio del producto",
            "example": 999.99
          },
          "stock": {
            "type": "integer",
            "description": "Cantidad en stock",
            "example": 50
          },
          "stockStatus": {
            "type": "string",
            "description": "Estado del stock calculado automaticamente",
            "enum": ["OK", "LOW", "OUT_OF_STOCK"],
            "example": "OK"
          }
        }
      },
...
    }
  }
}